feat: suspenseQuery 변경에 따른 대응
